% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_type_df}
\alias{cell_type_df}
\title{Example data set 2}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80 rows and 2 columns.
}
\usage{
cell_type_df
}
\description{
Example data set 2
}
\keyword{datasets}
