% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_abundance_sc_long}
\alias{get_abundance_sc_long}
\title{get abundance long}
\usage{
get_abundance_sc_long(.data, transcripts = NULL, all = F, exclude_zeros = F)
}
\arguments{
\item{.data}{A tidyseurat}

\item{transcripts}{A character}

\item{all}{A boolean}

\item{exclude_zeros}{A boolean}
}
\description{
get abundance long
}
\keyword{internal}
