% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{Example data set}
\format{
An object of class \code{Seurat} with 230 rows and 80 columns.
}
\usage{
pbmc_small
}
\description{
Example data set
}
\keyword{datasets}
