% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small_nested_interactions}
\alias{pbmc_small_nested_interactions}
\title{Example data set 2}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 9 columns.
}
\usage{
pbmc_small_nested_interactions
}
\description{
Example data set 2
}
\keyword{datasets}
