% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{return_args}
\alias{return_args}
\title{returns variables from an expression}
\usage{
return_args(args)
}
\arguments{
\item{args}{an expression}
}
\value{
list of symbols
}
\description{
returns variables from an expression
}
