% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{return_arguments_of}
\alias{return_arguments_of}
\title{returns variables from an expression}
\usage{
return_arguments_of(expression)
}
\arguments{
\item{expression}{an expression}
}
\value{
list of symbols
}
\description{
returns variables from an expression
}
