% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_method.R
\name{tbl_format_header}
\alias{tbl_format_header}
\title{Format the header of a tibble}
\description{
`r lifecycle::badge("experimental")`

For easier customization, the formatting of a tibble is split
into three components: header, body, and footer.
The `tbl_format_header()` method is responsible for formatting the header
of a tibble.

Override this method if you need to change the appearance
of the entire header.
If you only need to change or extend the components shown in the header,
override or extend [tbl_sum()] for your class which is called by the
default method.
}
