% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy}
\alias{tidy}
\alias{tidy.Seurat}
\title{tidy for Seurat objects}
\usage{
\method{tidy}{Seurat}(x, ...)
}
\arguments{
\item{x}{A Seurat object}

\item{...}{Additional arguments (not used)}
}
\value{
A tidyseurat object
}
\description{
tidy for Seurat objects
}
