% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_type_df}
\alias{cell_type_df}
\title{Cell types of 80 PBMC single cells}
\format{
A tibble containing 80 rows and 2 columns.
  Cells are a subsample of the Peripheral Blood Mononuclear Cells (PBMC) 
  dataset of 2,700 single cell. Cell types were identified with SingleR.
\describe{
  \item{cell}{cell identifier, barcode}
  \item{first.labels}{cell type}
}
}
\source{
\url{https://satijalab.org/seurat/v3.1/pbmc3k_tutorial.html}
}
\usage{
data(cell_type_df)
}
\value{
`tibble`
}
\description{
A dataset containing the barcodes and cell types of 80 PBMC single cells.
}
\keyword{datasets}
