% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_abundance_sc_wide}
\alias{get_abundance_sc_wide}
\title{get abundance wide}
\usage{
get_abundance_sc_wide(
  .data,
  features = NULL,
  all = FALSE,
  assay = .data@active.assay,
  slot = "data",
  prefix = ""
)
}
\arguments{
\item{.data}{A tidyseurat}

\item{features}{A character}

\item{all}{A boolean}

\item{assay}{assay name to extract feature abundance}

\item{slot}{slot in the assay, e.g. `data` and `scale.data`}

\item{prefix}{prefix for the feature names}
}
\value{
A Seurat object
}
\description{
get abundance wide
}
\examples{
data(pbmc_small)
pbmc_small \%>\%
  get_abundance_sc_wide(features=c("HLA-DRA", "LYZ"))

}
\keyword{internal}
