% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_abundance_sc_wide}
\alias{get_abundance_sc_wide}
\title{get abundance long}
\usage{
get_abundance_sc_wide(.data, transcripts = NULL, all = FALSE)
}
\arguments{
\item{.data}{A tidyseurat}

\item{transcripts}{A character}

\item{all}{A boolean}
}
\value{
A tidyseurat object
}
\description{
get abundance long
}
\keyword{internal}
