% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{join_transcripts}
\alias{join_transcripts}
\title{Extract and join information for transcripts.}
\usage{
join_transcripts(
  .data,
  transcripts = NULL,
  all = FALSE,
  exclude_zeros = FALSE,
  shape = "long"
)
}
\arguments{
\item{.data}{A tidyseurat object}

\item{transcripts}{A vector of transcript identifiers to join}

\item{all}{If TRUE return all}

\item{exclude_zeros}{If TRUE exclude zero values}

\item{shape}{Format of the returned table "long" or "wide"}
}
\value{
A `tbl` containing the information.for the specified transcripts
}
\description{
join_transcripts() extracts and joins information for specified transcripts
}
\details{
This function extracts information for specified transcripts and returns the information in either long or wide format.
}
\examples{

pbmc_small \%>\% 
tidy \%>\% 
join_transcripts(transcripts = c("HLA-DRA", "LYZ"))


}
