% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{tidy}
\alias{tidy}
\alias{tidy.Seurat}
\title{tidy for `Seurat`}
\usage{
tidy(object)

\method{tidy}{Seurat}(object)
}
\arguments{
\item{object}{A `Seurat` object.}
}
\value{
A `tidyseurat` object.
}
\description{
tidy for `Seurat`
}
\examples{
data(pbmc_small)
pbmc_small

}
